Fabio changes

+++ src/main/java/de/othr/im/gymbro/service/ExerciseService.java
@@ -63,7 +63,12 @@ public class ExerciseService {
     }
 
     public void deleteExercise(final Long id) {
-        this.exerciseRepository.deleteById(id);
+        if(this.getExercise(id) == null){
+            // TODO: Implement Warning
+            return;
+        } else {
+            this.exerciseRepository.deleteById(id);
+        }
     }
